Timer stretch goals
Possible glitch - when start button is not disabled, repeatedly pressing it will cause the timer to count far higher than anticipated.  This could be due to the previous interval still running, and them overlapping.  Cannot clear interval at the beginning of the timerFunction because it does not yet exist.  Will think on how to work around this, but timer works as expected with start and reset buttons.