Complete Tasks 2 and 3
Went through the index.html and use javascript to make it match the original.html.  There was a lot of repition is assigning values to selectors, so I experimented with iterating over a collection of elements in an attempt to dry up the code.  Ended up working really well and saved the headache of typing out similar things over and over.