Complete Task 4
I went through after the lecture and applied what we learned about adding elements to the page using Javascript.  Only way I could get it to work was by using querySelector on the parent to append and prepend the children.  getElementsByTagName didn't seem to work at all.  I couldn't apply the functions to the HTML Collection it seemed.