Complete stretch goals and first task for the digital timer
I finished the tasks for the project and went on to the two stretch goals.  Those two were pretty straight forward.  I then went on to attempt the timer.  That one was really tricky and I'm struggling with a few things about it.  Commiting what I have now before I begin a lot of experimenting.