Add reset button
I completed the SG3 task to add a functioning reset button.  The functionality for this button was much less complicated than the start button so it did not require much time.  It did require me to learn the :not() CSS selector to be able grab all the items in .digit class besides the one with the ID #colon.