Add style to the timer and buttons
I completed the final tasks for the stretch goals for this project.  The styling was simple.  I was surprised that I didn't forget anything after not using it much during the Javascript lessons.  Added comments to all of the js files in an attempt to get used to the idea of commenting code and to make things easier when I return to the code after a time.